// frontend/src/api/gen/schemas/accessControlMode.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * hackathon-portal-api
 * OpenAPI spec version: 1.0.0
 */

export type AccessControlMode =
  (typeof AccessControlMode)[keyof typeof AccessControlMode];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccessControlMode = {
  AuthenticationAuthorization: "AuthenticationAuthorization",
  Authentication: "Authentication",
  None: "None",
} as const;