// frontend/src/api/gen/schemas/managedIngressConfig.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * hackathon-portal-api
 * OpenAPI spec version: 1.0.0
 */
import type { AccessControlMode } from "./accessControlMode";

export interface ManagedIngressConfig {
  access_control_mode: AccessControlMode;
  /** @minimum 0 */
  server_port: number;
}